user weight displayed correctly + added functions to retrieve user info
Previously the user's weight would only be displayed once they updated it. If they already had a weight value in the database it would not be displayed. Added getUserField function in userFunctions.js that retrieves a specfied field of a user's document in our users database. Added useUser hook in firebaseFunctions.js that retrieves the user object of the current user if there is one signed in and returns null otherwise. This is Firebase's preferred method of getting current user instead of calling auth.currentUser so I will eventually migrate everything that currently uses getUser to use useUser instead. In UserInfo() added user variable and useEffect hook that triggers when user is modified. This is to fix a bug where user weight would not be displayed upon first loading of the page because useUser takes longer to run than loading the page does. This new hook ensures displaying the weight is not affected by timing issues like race conditions.