implement recipe search algorithm + fix DisplayMultipleRecipes()
There were timing issues with the previous code for searchRecipes() because it would return before promises were executed. Fixed this by making searchRecipes() async and used await instead of .then. Also fixed a bug where searchRecipes() was always returning an empty array because it wasn't ignoring the empty arrays in allLists.

DisplayMultipleRecipes didn't work previously with an actual array so I modified its implementation and renamed it DisplayRecipes(). Previously it used an html ordered list to display all the recipes, now it uses a variable that stores a mapping of every recipe to displayOneRecipe() called on that recipe. Also added some minor changes in the text that is displayed to the user.