increase search speed and fix spacing bug
Previously there was a bug with the recipe search algorithm where the first time it ran it would run quickly but all concurrent times it would run slowly. This was because for all concurrent times the recipes variable contained the contents of the previous search, which introduced overhead in rewriting each element. Now the reecipes variable is set to an empty array before every concurrent search.

The recipes were also displaying incorrectly, previously they were mushed together. Now there is a space divider that allows them to be displayed vertically in a line.