fix slow bmi calculation bug
The previous bmi algorithm had several nested async functions and re queried the user field objects every single time a search would be performed so it was extremely slow. The new algorithm retrieves the user's height and weight once and caches the values in local variables. The algorithm was completely redone and simplified to fewer functions and no longer uses async functions.